Expose logging config for calico cni
This change is exposing a new logging property in the installation object. With this change users are able to provide customized logging configuration for calico cni plugin via tigera-operator.

> updating api integer types to uint32
> adding validation for when cni is not calico but logging config is provided.

+++ pkg/render/node_test.go
@@ -54,8 +54,8 @@ var (
 	bgpDisabled          = operatorv1.BGPDisabled
 	nonPrivilegedEnabled = operatorv1.NonPrivilegedEnabled
 	logSeverity          = operatorv1.LogLevelDebug
-	logFileMaxAgeDays    = 5
-	logFileMaxCount      = 5
+	logFileMaxAgeDays    = uint32(5)
+	logFileMaxCount      = uint32(5)
 	logFileMaxSize       = resource.MustParse("1Mi")
 )
